refactor(status): one function per file
Split src/status.ts into one function per file per the new convention:
- status.ts          -> computeStatus (ctx + drift lookup inlined)
- status-classify.ts -> classify + DRIFT_STATES
- status-format.ts   -> formatStatusRow
- status-types.ts    -> StatusState / StatusRow / StatusOptions
No behaviour change; the same unit + integration tests pass.

skip-changelog: internal refactor, no consumer-observable surface
