+++ src/cli.ts
@@ -41,7 +41,8 @@ import { isAbsolute, resolve } from 'node:path';
 import { runChecks } from './check.js';
 import { plan } from './plan.js';
 import { publish } from './publish.js';
-import { computeStatus, formatStatusRow } from './status.js';
+import { formatStatusRow } from './status-format.js';
+import { computeStatus } from './status.js';
 import { writeCrateVersionForBuild } from './write-crate-version.js';
 import { writeLauncherFromConfig } from './write-launcher.js';
 import { writeVersionForBuild } from './write-version.js';
